Refactor: refine app and main func
