+++ bcs-services/bcs-webconsole/app/app.go
@@ -15,198 +15,215 @@ package app
 
 import (
 	"context"
-	"crypto/tls"
-	"fmt"
-	"os"
-	"path/filepath"
-	"strconv"
+	"net/http"
+	"os/signal"
+	"path"
 	"strings"
-	"time"
+	"syscall"
 
-	"github.com/Tencent/bk-bcs/bcs-common/common"
-	"github.com/Tencent/bk-bcs/bcs-common/common/blog"
-	comconf "github.com/Tencent/bk-bcs/bcs-common/common/conf"
 	"github.com/Tencent/bk-bcs/bcs-services/bcs-webconsole/app/options"
-	"github.com/Tencent/bk-bcs/bcs-services/bcs-webconsole/console/api"
-	"github.com/Tencent/bk-bcs/bcs-services/bcs-webconsole/console/config"
-	"github.com/Tencent/bk-bcs/bcs-services/bcs-webconsole/console/manager"
+	consoleConf "github.com/Tencent/bk-bcs/bcs-services/bcs-webconsole/console/config"
+	"github.com/Tencent/bk-bcs/bcs-services/bcs-webconsole/console/podmanager"
+	"github.com/Tencent/bk-bcs/bcs-services/bcs-webconsole/console/web"
+	"github.com/Tencent/bk-bcs/bcs-services/bcs-webconsole/handler"
+	"github.com/Tencent/bk-bcs/bcs-services/bcs-webconsole/i18n"
+	"github.com/Tencent/bk-bcs/bcs-services/bcs-webconsole/route"
+	yaml "github.com/asim/go-micro/plugins/config/encoder/yaml/v4"
 
-	"github.com/go-redis/redis/v8"
-	"k8s.io/apimachinery/pkg/util/wait"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd"
+	logger "github.com/Tencent/bk-bcs/bcs-common/common/blog"
+	"github.com/Tencent/bk-bcs/bcs-common/common/ssl"
+	etcd "github.com/asim/go-micro/plugins/registry/etcd/v4"
+	mhttp "github.com/asim/go-micro/plugins/server/http/v4"
+	"github.com/gin-contrib/cors"
+	"github.com/gin-gonic/gin"
+	"github.com/urfave/cli/v2"
+	"go-micro.dev/v4"
+	"go-micro.dev/v4/config"
+	"go-micro.dev/v4/config/reader"
+	"go-micro.dev/v4/config/reader/json"
+	"go-micro.dev/v4/config/source/file"
+	"go-micro.dev/v4/registry"
+	"golang.org/x/sync/errgroup"
 )
 
-// ConsoleManager is an console struct
-type ConsoleManager struct {
-	// options for console manager
-	opt *options.ConsoleManagerOption
+var (
+	// 变量, 编译后覆盖
+	service = "bcs-webconsole"
+	version = "latest"
+)
 
-	// tls config for cluster manager server
-	tlsConfig *tls.Config
-
-	k8sClient   *kubernetes.Clientset
-	k8sConfig   *rest.Config
-	redisClient *redis.Client // redis 客户端
-
-	backend manager.Manager
-	route   *api.Router
-	conf    *config.ConsoleConfig
+// WebConsoleManager is an console struct
+type WebConsoleManager struct {
+	ctx          context.Context
+	opt          *options.WebConsoleManagerOption
+	microService micro.Service
+	microConfig  config.Config
 }
 
-// NewConsoleManager create an ConsoleProxy object
-func NewConsoleManager(opt *options.ConsoleManagerOption) *ConsoleManager {
-
-	return &ConsoleManager{
-		opt:     opt,
-		backend: nil,
-		route:   nil,
-		conf:    nil,
+// NewWebConsoleManager
+func NewWebConsoleManager(opt *options.WebConsoleManagerOption) *WebConsoleManager {
+	return &WebConsoleManager{
+		ctx: context.Background(),
+		opt: opt,
 	}
 }
 
-func (c *ConsoleManager) Init() error {
+func (c *WebConsoleManager) Init() error {
+	// 初始化服务注册, 配置文件等
+	microService, microConfig := c.initMicroService()
+	c.microService = microService
+	c.microConfig = microConfig
 
-	err := setConfig(c.opt)
+	// etcd 服务发现注册
+	etcdRegistry, err := c.initEtcdRegistry()
 	if err != nil {
 		return err
 	}
-	// init server and client tls config
-	c.initTLSConfig()
-	// init redis
-	if err := c.initRedisCli(); err != nil {
+
+	if etcdRegistry != nil {
+		microService.Init(micro.Registry(etcdRegistry))
+	}
+
+	// http 路由注册
+	router, err := c.initHTTPService()
+	if err != nil {
 		return err
 	}
-	// init k8s client
-	if err := c.initK8sClient(); err != nil {
+
+	if err := micro.RegisterHandler(microService.Server(), router); err != nil {
 		return err
 	}
 
 	return nil
+}
+
+func (c *WebConsoleManager) initMicroService() (micro.Service, config.Config) {
+	var configPath string
+
+	// new config
+	conf, _ := config.NewConfig(
+		config.WithReader(json.NewReader(reader.WithEncoder(yaml.NewEncoder()))),
+	)
+
+	confFlags := micro.Flags(
+		&cli.StringFlag{
+			Name:        "bcs-conf",
+			Usage:       "bcs-config path",
+			Required:    true,
+			Destination: &configPath,
+		},
+	)
+
+	confAction := micro.Action(func(c *cli.Context) error {
+		logger.Info("load conf from ", configPath)
+		if err := conf.Load(file.NewSource(file.WithPath(configPath))); err != nil {
+			return err
+		}
+
+		// 初始化配置文件
+		consoleConf.G.ReadFrom(conf.Bytes())
+
+		return nil
+	})
+
+	srv := micro.NewService(micro.Server(mhttp.NewServer()), confFlags)
+	opts := []micro.Option{
+		micro.Name(service),
+		micro.Version(version),
+		confAction,
+	}
+
+	// 配置文件, 日志这里才设置完成
+	srv.Init(opts...)
+
+	return srv, conf
+}
+
+// initHTTPService 初始化 gin Http 配置
+func (c *WebConsoleManager) initHTTPService() (*gin.Engine, error) {
+	gin.SetMode(gin.ReleaseMode)
+	router := gin.New()
+	router.Use(gin.Recovery(), gin.Logger(), cors.Default())
+	router.Use(i18n.Localize())
+
+	// 注册模板和静态资源
+	router.SetHTMLTemplate(web.WebTemplate())
+
+	// 静态资源
+	routePrefix := consoleConf.G.Web.RoutePrefix
+	if routePrefix == "" {
+		routePrefix = "/" + service
+	}
+
+	// 支持路径 prefix 透传和 rewrite 的场景
+	router.StaticFS(path.Join(routePrefix, "/web/static"), http.FS(web.WebStatic()))
+	router.StaticFS("/web/static", http.FS(web.WebStatic()))
+
+	handlerOpts := &route.Options{
+		RoutePrefix: routePrefix,
+		Client:      c.microService.Client(),
+		Router:      router,
+	}
+
+	if err := handler.Register(handlerOpts); err != nil {
+		return nil, err
+	}
+
+	return router, nil
+}
+
+// initEtcdRegistry etcd 服务注册
+func (c *WebConsoleManager) initEtcdRegistry() (registry.Registry, error) {
+	ca := c.microConfig.Get("etcd", "ca").String("")
+	cert := c.microConfig.Get("etcd", "cert").String("")
+	key := c.microConfig.Get("etcd", "key").String("")
+	if ca == "" || cert == "" || key == "" {
+		return nil, nil
+	}
+
+	endpoints := c.microConfig.Get("etcd", "endpoints").String("127.0.0.1:2379")
+	etcdRegistry := etcd.NewRegistry(registry.Addrs(strings.Split(endpoints, ",")...))
+	tlsConfig, err := ssl.ClientTslConfVerity(ca, cert, key, "")
+	if err != nil {
+		return nil, err
+	}
+	etcdRegistry.Init(registry.TLSConfig(tlsConfig))
+
+	return etcdRegistry, nil
 }
 
 // Run create a pid
-func (c *ConsoleManager) Run() error {
+func (c *WebConsoleManager) Run() error {
+	// Create context that listens for the interrupt signal from the OS.
+	ctx, stop := signal.NotifyContext(c.ctx, syscall.SIGINT, syscall.SIGTERM)
+	defer stop()
 
-	//pid
-	if err := common.SavePid(comconf.ProcessConfig{}); err != nil {
-		blog.Error("fail to save pid: err:%s", err.Error())
-	}
+	logger.Info("starting bcs-webconsole.")
 
-	c.backend = manager.NewManager(&c.opt.Conf, c.k8sClient, c.k8sConfig, c.redisClient, nil)
-	c.route = api.NewRouter(c.backend, &c.opt.Conf)
-	stopCh := make(chan struct{})
-	// 定期清理pod
-	go wait.NonSlidingUntil(c.backend.CleanUserPod, manager.CleanUserPodInterval*time.Second, stopCh)
+	c.microService.Init(micro.AfterStop(func() error {
+		// 会让 websocket 发送 EndOfTransmission, 不能保证一定发送成功
+		logger.Info("receive interput, gracefully shutdown")
+		<-ctx.Done()
+		return nil
+	}))
 
-	return nil
-}
+	eg, ctx := errgroup.WithContext(ctx)
 
-func (c *ConsoleManager) initTLSConfig() {
+	podCleanUpMgr := podmanager.NewCleanUpManager(ctx)
+	eg.Go(func() error {
+		return podCleanUpMgr.Run()
+	})
 
-	// server cert directoty
-	if c.opt.CertConfig.ServerCertFile != "" && c.opt.CertConfig.CAFile != "" && c.opt.CertConfig.ServerKeyFile != "" {
-		c.opt.Conf.ServCert.IsSSL = true
-	}
-}
-
-func (c *ConsoleManager) initRedisCli() error {
-
-	dbNum, err := strconv.Atoi(c.opt.Redis.Database)
-	if nil != err {
-		return err
-	}
-	if c.opt.Redis.PoolSize == 0 {
-		c.opt.Redis.PoolSize = 3000
-	}
-
-	var client *redis.Client
-
-	if c.opt.Redis.MasterName == "" {
-		option := &redis.Options{
-			Addr:     c.opt.Redis.Address,
-			Password: c.opt.Redis.Password,
-			DB:       dbNum,
-			PoolSize: c.opt.Redis.PoolSize,
-		}
-		client = redis.NewClient(option)
-
-	} else {
-		hosts := strings.Split(c.opt.Redis.Address, ",")
-		option := &redis.FailoverOptions{
-			MasterName:       c.opt.Redis.MasterName,
-			SentinelAddrs:    hosts,
-			Password:         c.opt.Redis.Password,
-			DB:               dbNum,
-			PoolSize:         c.opt.Redis.PoolSize,
-			SentinelPassword: c.opt.Redis.SentinelPassword,
-		}
-		client = redis.NewFailoverClient(option)
-	}
-
-	err = client.Ping(context.Background()).Err()
-	if err != nil {
-		return err
-	}
-
-	c.redisClient = client
-
-	return nil
-}
-
-func (c *ConsoleManager) initK8sClient() error {
-	// 配置 k8s 集群外 kubeconfig 配置文件
-	if home := homeDir(); home != "" {
-		c.opt.KubeConfigFile = filepath.Join(home, ".kube", "config")
-	}
-
-	//在 kubeconfig 中使用当前上下文环境，config 获取支持 url 和 path 方式
-	k8sConfig, err := rest.InClusterConfig()
-	if err != nil {
-		k8sConfig, err = clientcmd.BuildConfigFromFlags("", c.opt.KubeConfigFile)
-		if err != nil {
+	eg.Go(func() error {
+		if err := c.microService.Run(); err != nil {
 			return err
 		}
-	}
+		return nil
+	})
 
-	//在 kubeconfig 中使用当前上下文环境，config 获取支持 url 和 path 方式
-	config, err := clientcmd.BuildConfigFromFlags("", c.opt.KubeConfigFile)
-	if err != nil {
+	if err := eg.Wait(); err != nil {
+		defer logger.CloseLogs()
 		return err
 	}
-
-	k8sClient, err := kubernetes.NewForConfig(config)
-	if err != nil {
-		return err
-	}
-
-	c.k8sConfig = k8sConfig
-	c.k8sClient = k8sClient
-
 	return nil
-}
-
-func setConfig(op *options.ConsoleManagerOption) error {
-	op.Redis.Address = op.RedisAddress
-	op.Redis.Password = op.RedisPassword
-	op.Redis.Database = op.RedisDatabase
-	op.Redis.MasterName = op.RedisMasterName
-	op.Redis.SentinelPassword = op.RedisSentinelPassword
-	op.Redis.PoolSize = op.RedisPoolSize
-
-	op.Conf.Address = op.Address
-	op.Conf.Port = int(op.Port)
-	if op.WebConsoleImage == "" {
-		return fmt.Errorf("web-console-image required")
-	}
-	op.Conf.WebConsoleImage = op.WebConsoleImage
-	return nil
-}
-
-func homeDir() string {
-	if h := os.Getenv("HOME"); h != "" {
-		return h
-	}
-	return os.Getenv("USERPROFILE") // windows
 }
